Further improvements to ProtocolMetadata
