+++ src/main/java/net/tridentsdk/server/entity/block/TridentArmorStand.java
@@ -23,14 +23,13 @@ import net.tridentsdk.entity.block.ArmorStand;
 import net.tridentsdk.entity.block.SlotProperties;
 import net.tridentsdk.entity.living.Player;
 import net.tridentsdk.event.entity.EntityDamageEvent;
+import net.tridentsdk.server.data.MetadataType;
 import net.tridentsdk.server.data.ProtocolMetadata;
 import net.tridentsdk.server.entity.TridentLivingEntity;
 import net.tridentsdk.util.PartRotation;
 import net.tridentsdk.window.inventory.Item;
 
 import java.util.UUID;
-
-import static net.tridentsdk.server.data.ProtocolMetadata.MetadataType;
 
 public class TridentArmorStand extends TridentLivingEntity implements ArmorStand {
     private final SlotProperties properties;
