+++ src/main/java/net/tridentsdk/server/entity/living/TridentPig.java
@@ -24,6 +24,7 @@ import net.tridentsdk.entity.Projectile;
 import net.tridentsdk.entity.living.Pig;
 import net.tridentsdk.entity.living.Player;
 import net.tridentsdk.event.entity.EntityDamageEvent;
+import net.tridentsdk.server.data.MetadataType;
 import net.tridentsdk.server.data.ProtocolMetadata;
 import net.tridentsdk.server.entity.TridentAgeable;
 import net.tridentsdk.server.entity.TridentLivingEntity;
@@ -42,7 +43,7 @@ public class TridentPig extends TridentAgeable implements Pig {
     protected void encodeMetadata(ProtocolMetadata protocolMeta) {
         super.encodeMetadata(protocolMeta);
 
-        protocolMeta.setMeta(16, ProtocolMetadata.MetadataType.BYTE, (hasSaddle) ? (byte) 1 : (byte) 0);
+        protocolMeta.setMeta(16, MetadataType.BYTE, (hasSaddle) ? (byte) 1 : (byte) 0);
     }
 
     @Override
